+++ tgbotapi.extensions.behaviour_builder/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/behaviour_builder/BehaviourBuilders.kt
@@ -11,6 +11,42 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.plus
 
 /**
+ * This function is used in [buildBehaviour] extensions to provide default [CoroutineScope] and allow to avoid all
+ * unnecessary parameters except of block
+ */
+expect var defaultCoroutineScopeProvider: () -> CoroutineScope
+
+/**
+ * Use this method in case you wish to make some additional actions with [flowUpdatesFilter].
+ *
+ * **WARNING** This method WILL NOT launch any listening of updates. Use something like
+ * [startGettingOfUpdatesByLongPolling] or tools for work with webhooks
+ *
+ * @see [BehaviourContext]
+ * @see startGettingOfUpdatesByLongPolling
+ */
+@PreviewFeature
+@Deprecated("Parameters has been reordered. Replace scope and flowUpdatesFilter for correct order")
+suspend fun TelegramBot.buildBehaviour(
+    scope: CoroutineScope,
+    flowUpdatesFilter: FlowsUpdatesFilter,
+    defaultExceptionsHandler: ExceptionHandler<Unit>? = null,
+    block: BehaviourContextReceiver<Unit>
+) {
+    BehaviourContext(
+        this,
+        scope.let {
+              if (defaultExceptionsHandler == null) {
+                  it
+              } else {
+                  it + ContextSafelyExceptionHandler(defaultExceptionsHandler)
+              }
+        },
+        flowUpdatesFilter
+    ).block()
+}
+
+/**
  * Use this method in case you wish to make some additional actions with [flowUpdatesFilter].
  *
  * **WARNING** This method WILL NOT launch any listening of updates. Use something like
@@ -21,8 +57,8 @@ import kotlinx.coroutines.plus
  */
 @PreviewFeature
 suspend fun TelegramBot.buildBehaviour(
-    scope: CoroutineScope,
     flowUpdatesFilter: FlowsUpdatesFilter,
+    scope: CoroutineScope = defaultCoroutineScopeProvider(),
     defaultExceptionsHandler: ExceptionHandler<Unit>? = null,
     block: BehaviourContextReceiver<Unit>
 ) {
@@ -49,7 +85,7 @@ suspend fun TelegramBot.buildBehaviour(
  */
 @PreviewFeature
 suspend fun TelegramBot.buildBehaviour(
-    scope: CoroutineScope,
+    scope: CoroutineScope = defaultCoroutineScopeProvider(),
     defaultExceptionsHandler: ExceptionHandler<Unit>? = null,
     block: BehaviourContextReceiver<Unit>
 ) = FlowsUpdatesFilter().let {